requirements.txt
asgiref==3.4.1
backcall==0.2.0
binary==1.0.0
debugpy==1.4.1
decorator==5.0.9
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-sites==0.10
entrypoints==0.3
gunicorn==20.1.0
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==7.0.1
jupyter-core==4.7.1
matplotlib-inline==0.1.2
nest-asyncio==1.5.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.2.1
six==1.16.0
sqlparse==0.4.2
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
